+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## 10.0.3 (unreleased)
+## 10.0.6 (unreleased)
 
 ### Added
 
@@ -6,7 +6,27 @@
 
 ### Changed
 
-### Remved
+### Removed
+
+## 10.0.5
+### Fixed
+- Fixed captureBothSides on iOS being inverted
+
+
+## 10.0.4
+
+### Fixed
+- Fixed missing selfie instructions screen on DocV and EnhancedDocV
+- Fixed broken disable capture both sides flag
+- Fixed show skip button on capture both sides
+
+## 10.0.3
+
+### Added
+- Added allowNewEnroll on SmartSelfie, BiometricKYC, DocV and EnhancedDocV
+
+### Fixed
+- Fixed missing callbackUrl  
 
 ## 10.0.2